# web/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Robot Operation Center (ROC). Internal automation platform to distribute and execute robots across different subnets.",
    "version": "1.0.0",
    "title": "ROC (Robot Operation Center)",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/task-force": {
      "get": {
        "operationId": "getTaskForces",
        "summary": "List Task Forces",
        "tags": [
          "task-force"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskForce"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTaskForce",
        "summary": "Create new task force",
        "tags": [
          "task-force"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskForce"
                }
              }
            }
          }
        }
      }
    },
    "/task-force/{taskForceId}": {
      "get": {
        "operationId": "getTaskForceById",
        "summary": "Get Task Force",
        "tags": [
          "task-force"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskForceId",
            "description": "Id of task force",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        }
      },
      "put": {
        "operationId": "editTaskForce",
        "summary": "Edit task force",
        "tags": [
          "task-force"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskForceId",
            "description": "Id of task force",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        }
      },
      "delete": {
        "operationId": "deleteTaskForce",
        "summary": "Deletes a task force",
        "tags": [
          "task-force"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskForceId",
            "description": "Id of task force",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Token",
        "description": "Provide access token to perform operations."
      }
    },
...
    "schemas": {
...
      "TaskForce": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "runner": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "environmentId": {
            "type": "string"
          },
          "tasks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "example": {
          "id": "054fb42",
          "name": "Daily Health Checks",
          "runner": "gcr.io/yusufcanb/roc-runner:latest",
          "agentId": "1",
          "environmentId": "1",
          "tasks": [
            "demo.robot"
          ]
        }
      },
...
    }
  }
}